+++ src/DCC/MainRoutines.cpp
@@ -358,7 +358,7 @@ uint8_t DCC::readCVBytesMain(uint16_t addr, uint16_t cv,
 
 }
 
-uint8_t DCC::getThrottleSpeed(uint8_t cab) {
+uint8_t DCC::getThrottleSpeed(uint16_t cab) {
   int reg=lookupSpeedTable(cab);
   if (reg < 0) {
     return -1;
@@ -366,7 +366,7 @@ uint8_t DCC::getThrottleSpeed(uint8_t cab) {
   return speedTable[reg].speedCode & 0x7F;
 }
 
-bool DCC::getThrottleDirection(uint8_t cab) {
+bool DCC::getThrottleDirection(uint16_t cab) {
   int reg=lookupSpeedTable(cab);
   if (reg < 0) {
     return false;
@@ -379,7 +379,7 @@ uint8_t DCC::speedAndDirToCode(uint8_t speed, bool dir) {
   return (speed & 0x7F) + (dir ? 128 : 0); 
 } 
 
-void DCC::updateSpeedTable(uint8_t cab, uint8_t speedCode) {
+void DCC::updateSpeedTable(uint16_t cab, uint8_t speedCode) {
   if(cab == 0) {
     // broadcast to all locomotives
     for(int dev = 0; dev < numDevices; dev++) 
@@ -391,7 +391,7 @@ void DCC::updateSpeedTable(uint8_t cab, uint8_t speedCode) {
   if(reg >= 0) speedTable[reg].speedCode = speedCode;
 }
 
-int DCC::lookupSpeedTable(uint8_t cab) {
+int DCC::lookupSpeedTable(uint16_t cab) {
   int firstEmpty = numDevices;
   int reg;
   for(reg = 0; reg < numDevices; reg++) {
@@ -412,7 +412,7 @@ int DCC::lookupSpeedTable(uint8_t cab) {
   return reg;
 }
 
-void DCC::forgetDevice(uint8_t cab) {  // removes any speed reminders for this loco  
+void DCC::forgetDevice(uint16_t cab) {  // removes any speed reminders for this loco  
   int reg = lookupSpeedTable(cab);
 
   if(reg >= 0) speedTable[reg].cab = 0;
